requirements.txt
pandas
numpy
spotipy
flask
slackclient
gunicorn